Factor test helpers into discrete modules
