+++ test/bandit/http2/protocol_test.exs
@@ -1,13 +1,12 @@
 defmodule HTTP2ProtocolTest do
-  use ConnectionHelpers, async: true
-
-  import Plug.Conn
+  use ExUnit.Case, async: true
+  use ServerHelpers
 
   setup :https_server
 
   describe "frame splitting / merging" do
     test "it should handle cases where the request arrives in small chunks", context do
-      socket = tls_client(context)
+      socket = SimpleH2Client.tls_client(context)
 
       # Send connection preface, client settings & ping frame one byte at a time
       ("PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n" <>
@@ -24,7 +23,7 @@ defmodule HTTP2ProtocolTest do
     end
 
     test "it should handle cases where multiple frames arrive in the same packet", context do
-      socket = tls_client(context)
+      socket = SimpleH2Client.tls_client(context)
 
       # Send connection preface, client settings & ping frame all in one
       :ssl.send(
@@ -42,16 +41,16 @@ defmodule HTTP2ProtocolTest do
   describe "errors and unexpected frames" do
     @tag capture_log: true
     test "it should ignore unknown frame types", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 0, 254, 0, 0, 0, 0, 0>>)
-      assert connection_alive?(socket)
+      assert SimpleH2Client.connection_alive?(socket)
     end
 
     @tag capture_log: true
     test "it should shut down the connection gracefully when encountering a connection error",
          context do
-      socket = tls_client(context)
-      exchange_prefaces(socket)
+      socket = SimpleH2Client.tls_client(context)
+      SimpleH2Client.exchange_prefaces(socket)
       # Send a bogus SETTINGS frame
       :ssl.send(socket, <<0, 0, 0, 4, 0, 0, 0, 0, 1>>)
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1>>}
@@ -61,13 +60,13 @@ defmodule HTTP2ProtocolTest do
   describe "connection preface handling" do
     @tag capture_log: true
     test "closes with an error if the HTTP/2 connection preface is not present", context do
-      socket = tls_client(context)
+      socket = SimpleH2Client.tls_client(context)
       :ssl.send(socket, "PRI * NOPE/2.0\r\n\r\nSM\r\n\r\n")
       assert :ssl.recv(socket, 0) == {:error, :closed}
     end
 
     test "the server should send a SETTINGS frame at start of the connection", context do
-      socket = tls_client(context)
+      socket = SimpleH2Client.tls_client(context)
       :ssl.send(socket, "PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n")
       assert :ssl.recv(socket, 9) == {:ok, <<0, 0, 0, 4, 0, 0, 0, 0, 0>>}
     end
@@ -75,17 +74,17 @@ defmodule HTTP2ProtocolTest do
 
   describe "DATA frames" do
     test "sends end of stream when there is a single data frame", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "GET"},
         {":path", "/body_response"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      assert successful_response?(socket, 1, false)
-      assert simple_read_body(socket) == {:ok, 1, true, "OK"}
+      assert SimpleH2Client.successful_response?(socket, 1, false)
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, "OK"}
     end
 
     def body_response(conn) do
@@ -93,19 +92,19 @@ defmodule HTTP2ProtocolTest do
     end
 
     test "sends multiple DATA frames with last one end of stream when chunking", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "GET"},
         {":path", "/chunk_response"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      assert successful_response?(socket, 1, false)
-      assert simple_read_body(socket) == {:ok, 1, false, "OK"}
-      assert simple_read_body(socket) == {:ok, 1, false, "DOKEE"}
-      assert simple_read_body(socket) == {:ok, 1, true, ""}
+      assert SimpleH2Client.successful_response?(socket, 1, false)
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, false, "OK"}
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, false, "DOKEE"}
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, ""}
     end
 
     def chunk_response(conn) do
@@ -118,9 +117,9 @@ defmodule HTTP2ProtocolTest do
     end
 
     test "reads a zero byte body if none is sent", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "POST"},
         {":path", "/echo"},
         {":scheme", "https"},
@@ -128,7 +127,7 @@ defmodule HTTP2ProtocolTest do
       ])
 
       # A zero byte body being written will cause end_stream to be set on the header frame
-      assert successful_response?(socket, 1, true)
+      assert SimpleH2Client.successful_response?(socket, 1, true)
     end
 
     def echo(conn) do
@@ -138,91 +137,91 @@ defmodule HTTP2ProtocolTest do
 
     @tag capture_log: true
     test "rejects DATA frames received on an idle stream", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_body(socket, 1, true, "OK")
+      SimpleH2Client.send_body(socket, 1, true, "OK")
 
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1>>}
     end
 
     test "reads a one frame body if one frame is sent", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, false, [
+      SimpleH2Client.send_headers(socket, 1, false, [
         {":method", "POST"},
         {":path", "/echo"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      simple_send_body(socket, 1, true, "OK")
+      SimpleH2Client.send_body(socket, 1, true, "OK")
 
-      assert successful_response?(socket, 1, false)
-      assert simple_read_body(socket) == {:ok, 1, true, "OK"}
+      assert SimpleH2Client.successful_response?(socket, 1, false)
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, "OK"}
     end
 
     test "reads a multi frame body if many frames are sent", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, false, [
+      SimpleH2Client.send_headers(socket, 1, false, [
         {":method", "POST"},
         {":path", "/echo"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      simple_send_body(socket, 1, false, "OK")
-      simple_send_body(socket, 1, true, "OK")
+      SimpleH2Client.send_body(socket, 1, false, "OK")
+      SimpleH2Client.send_body(socket, 1, true, "OK")
 
-      assert successful_response?(socket, 1, false)
-      assert simple_read_body(socket) == {:ok, 1, true, "OKOK"}
+      assert SimpleH2Client.successful_response?(socket, 1, false)
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, "OKOK"}
     end
 
     @tag capture_log: true
     test "rejects DATA frames received on a remote closed stream", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "POST"},
         {":path", "/echo"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      simple_send_body(socket, 1, true, "OK")
+      SimpleH2Client.send_body(socket, 1, true, "OK")
 
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 1>>}
     end
 
     @tag capture_log: true
     test "rejects DATA frames received on a zero stream id", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_body(socket, 0, true, "OK")
+      SimpleH2Client.send_body(socket, 0, true, "OK")
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1>>}
     end
 
     @tag capture_log: true
     test "rejects DATA frames received on an invalid stream id", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_body(socket, 2, true, "OK")
+      SimpleH2Client.send_body(socket, 2, true, "OK")
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1>>}
     end
   end
 
   describe "HEADERS frames" do
     test "sends end of stream headers when there is no body", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "GET"},
         {":path", "/no_body_response"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      assert simple_read_headers(socket) ==
+      assert SimpleH2Client.read_headers(socket) ==
                {:ok, 1, true,
                 [{":status", "200"}, {"cache-control", "max-age=0, private, must-revalidate"}]}
     end
@@ -232,38 +231,38 @@ defmodule HTTP2ProtocolTest do
     end
 
     test "sends non-end of stream headers when there is a body", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "GET"},
         {":path", "/body_response"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      assert simple_read_headers(socket) ==
+      assert SimpleH2Client.read_headers(socket) ==
                {:ok, 1, false,
                 [{":status", "200"}, {"cache-control", "max-age=0, private, must-revalidate"}]}
 
-      assert simple_read_body(socket) == {:ok, 1, true, "OK"}
+      assert(SimpleH2Client.read_body(socket) == {:ok, 1, true, "OK"})
     end
 
     test "accepts well-formed headers without padding or priority", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       headers = headers_for_header_read_test(context)
 
       # Send unadorned headers
       :ssl.send(socket, [<<0, 0, byte_size(headers), 1, 0x05, 0, 0, 0, 1>>, headers])
 
-      assert simple_read_headers(socket) ==
+      assert SimpleH2Client.read_headers(socket) ==
                {:ok, 1, false,
                 [{":status", "200"}, {"cache-control", "max-age=0, private, must-revalidate"}]}
 
-      assert simple_read_body(socket) == {:ok, 1, true, "OK"}
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, "OK"}
     end
 
     test "accepts well-formed headers with priority", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       headers = headers_for_header_read_test(context)
 
       # Send headers with priority
@@ -273,15 +272,15 @@ defmodule HTTP2ProtocolTest do
         headers
       ])
 
-      assert simple_read_headers(socket) ==
+      assert SimpleH2Client.read_headers(socket) ==
                {:ok, 1, false,
                 [{":status", "200"}, {"cache-control", "max-age=0, private, must-revalidate"}]}
 
-      assert simple_read_body(socket) == {:ok, 1, true, "OK"}
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, "OK"}
     end
 
     test "accepts well-formed headers with padding", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       headers = headers_for_header_read_test(context)
 
       # Send headers with padding
@@ -292,15 +291,15 @@ defmodule HTTP2ProtocolTest do
         <<1, 2, 3, 4>>
       ])
 
-      assert simple_read_headers(socket) ==
+      assert SimpleH2Client.read_headers(socket) ==
                {:ok, 1, false,
                 [{":status", "200"}, {"cache-control", "max-age=0, private, must-revalidate"}]}
 
-      assert simple_read_body(socket) == {:ok, 1, true, "OK"}
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, "OK"}
     end
 
     test "accepts well-formed headers with padding and priority", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       headers = headers_for_header_read_test(context)
 
       # Send headers with padding and priority
@@ -311,11 +310,11 @@ defmodule HTTP2ProtocolTest do
         <<1, 2, 3, 4>>
       ])
 
-      assert simple_read_headers(socket) ==
+      assert SimpleH2Client.read_headers(socket) ==
                {:ok, 1, false,
                 [{":status", "200"}, {"cache-control", "max-age=0, private, must-revalidate"}]}
 
-      assert simple_read_body(socket) == {:ok, 1, true, "OK"}
+      assert SimpleH2Client.read_body(socket) == {:ok, 1, true, "OK"}
     end
 
     def headers_for_header_read_test(context) do
@@ -341,7 +340,7 @@ defmodule HTTP2ProtocolTest do
     @tag capture_log: true
     test "closes with an error when receiving a zero stream ID",
          context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 5, 1, 0x04, 0, 0, 0, 0, 64, 129, 31, 129, 31>>)
 
       assert :ssl.recv(socket, 17) ==
@@ -353,7 +352,7 @@ defmodule HTTP2ProtocolTest do
     @tag capture_log: true
     test "closes with an error when receiving an even stream ID",
          context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 5, 1, 0x04, 0, 0, 0, 98, 64, 129, 31, 129, 31>>)
 
       assert :ssl.recv(socket, 17) ==
@@ -365,7 +364,7 @@ defmodule HTTP2ProtocolTest do
     @tag capture_log: true
     test "closes with an error when receiving a stream ID we've already seen",
          context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 5, 1, 0x04, 0, 0, 0, 99, 64, 129, 31, 129, 31>>)
       :ssl.send(socket, <<0, 0, 5, 1, 0x04, 0, 0, 0, 99, 64, 129, 31, 129, 31>>)
 
@@ -377,7 +376,7 @@ defmodule HTTP2ProtocolTest do
 
     @tag capture_log: true
     test "closes with an error on a header frame with undecompressable header block", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 11, 1, 0x2C, 0, 0, 0, 1, 2, 1::1, 12::31, 34, 1, 2, 3, 4, 5>>)
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 9>>}
       assert :ssl.recv(socket, 0) == {:error, :closed}
@@ -387,39 +386,39 @@ defmodule HTTP2ProtocolTest do
   describe "RST_STREAM frames" do
     @tag capture_log: true
     test "sends RST_FRAME with no error if stream task ends without closed stream", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
       # Send headers with end_stream bit cleared
-      simple_send_headers(socket, 1, false, [
+      SimpleH2Client.send_headers(socket, 1, false, [
         {":method", "GET"},
         {":path", "/body_response"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      simple_read_headers(socket)
-      simple_read_body(socket)
+      SimpleH2Client.read_headers(socket)
+      SimpleH2Client.read_body(socket)
 
-      assert read_rst_stream(socket) == {:ok, 1, 0}
-      assert connection_alive?(socket)
+      assert SimpleH2Client.read_rst_stream(socket) == {:ok, 1, 0}
+      assert SimpleH2Client.connection_alive?(socket)
     end
 
     @tag capture_log: true
     test "sends RST_FRAME with error if stream task crashes", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "GET"},
         {":path", "/crasher"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      simple_read_headers(socket)
-      simple_read_body(socket)
+      SimpleH2Client.read_headers(socket)
+      SimpleH2Client.read_body(socket)
 
-      assert read_rst_stream(socket) == {:ok, 1, 2}
-      assert connection_alive?(socket)
+      assert SimpleH2Client.read_rst_stream(socket) == {:ok, 1, 2}
+      assert SimpleH2Client.connection_alive?(socket)
     end
 
     def crasher(conn) do
@@ -432,34 +431,34 @@ defmodule HTTP2ProtocolTest do
 
     @tag capture_log: true
     test "rejects RST_STREAM frames received on an idle stream", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      send_rst_stream(socket, 1, 0)
+      SimpleH2Client.send_rst_stream(socket, 1, 0)
 
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1>>}
     end
 
     test "shuts down the stream task on receipt of an RST_STREAM frame", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      simple_send_headers(socket, 1, true, [
+      SimpleH2Client.send_headers(socket, 1, true, [
         {":method", "GET"},
         {":path", "/sleeper"},
         {":scheme", "https"},
         {":authority", "localhost:#{context.port}"}
       ])
 
-      simple_read_headers(socket)
-      {:ok, 1, false, "OK"} = simple_read_body(socket)
+      SimpleH2Client.read_headers(socket)
+      {:ok, 1, false, "OK"} = SimpleH2Client.read_body(socket)
 
       assert Process.whereis(:sleeper) |> Process.alive?()
 
-      send_rst_stream(socket, 1, 0)
+      SimpleH2Client.send_rst_stream(socket, 1, 0)
 
       Process.sleep(100)
 
       assert Process.whereis(:sleeper) == nil
-      assert connection_alive?(socket)
+      assert SimpleH2Client.connection_alive?(socket)
     end
 
     def sleeper(conn) do
@@ -475,8 +474,8 @@ defmodule HTTP2ProtocolTest do
 
   describe "SETTINGS frames" do
     test "the server should acknowledge a client's SETTINGS frames", context do
-      socket = tls_client(context)
-      exchange_prefaces(socket)
+      socket = SimpleH2Client.tls_client(context)
+      SimpleH2Client.exchange_prefaces(socket)
       :ssl.send(socket, <<0, 0, 0, 4, 0, 0, 0, 0, 0>>)
       assert :ssl.recv(socket, 9) == {:ok, <<0, 0, 0, 4, 1, 0, 0, 0, 0>>}
     end
@@ -484,7 +483,7 @@ defmodule HTTP2ProtocolTest do
 
   describe "PING frames" do
     test "the server should acknowledge a client's PING frames", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 8, 6, 0, 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8>>)
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 6, 1, 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8>>}
     end
@@ -492,9 +491,9 @@ defmodule HTTP2ProtocolTest do
 
   describe "GOAWAY frames" do
     test "the server should send a GOAWAY frame when shutting down", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
 
-      assert connection_alive?(socket)
+      assert SimpleH2Client.connection_alive?(socket)
 
       Process.sleep(100)
 
@@ -505,14 +504,14 @@ defmodule HTTP2ProtocolTest do
     end
 
     test "the server should close the connection upon receipt of a GOAWAY frame", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>)
       assert :ssl.recv(socket, 17) == {:ok, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>}
       assert :ssl.recv(socket, 0) == {:error, :closed}
     end
 
     test "the server should return the last received stream id in the GOAWAY frame", context do
-      socket = setup_connection(context)
+      socket = SimpleH2Client.setup_connection(context)
       :ssl.send(socket, <<0, 0, 5, 1, 0x04, 0, 0, 0, 99, 64, 129, 31, 129, 31>>)
       :ssl.send(socket, <<0, 0, 8, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>)
