+++ test/bandit/initial_handler_test.exs
@@ -1,7 +1,7 @@
 defmodule InitialHandlerTest do
-  use ConnectionHelpers, async: true
-
-  import Plug.Conn
+  use ExUnit.Case, async: true
+  use ServerHelpers
+  use FinchHelpers
 
   def report_version(conn) do
     send_resp(conn, 200, conn |> get_http_protocol() |> to_string())
@@ -9,7 +9,7 @@ defmodule InitialHandlerTest do
 
   describe "HTTP 1.x handling" do
     setup :http_server
-    setup :http1_client
+    setup :finch_http1_client
 
     test "sets up the HTTP 1.x handler", %{base: base, finch_name: finch_name} do
       {:ok, response} = Finch.build(:get, base <> "/report_version") |> Finch.request(finch_name)
@@ -21,7 +21,7 @@ defmodule InitialHandlerTest do
 
   describe "HTTP/2 handling" do
     setup :https_server
-    setup :http2_client
+    setup :finch_h2_client
 
     test "sets up the HTTP/2 handler", %{base: base, finch_name: finch_name} do
       {:ok, response} = Finch.build(:get, base <> "/report_version") |> Finch.request(finch_name)
